Fixed a JPEG exporter bug and a possible problem in the little endian to array writer, regarding signedness of 0xFF000000, reported by Christoph.

+++ src/util.h
@@ -6,19 +6,19 @@
 	array[ index   ] =  value & 0x000000FF;\
 	array[ index+1 ] = (value & 0x0000FF00)>>8;\
 	array[ index+2 ] = (value & 0x00FF0000)>>16;\
-	array[ index+3 ] = (value & 0xFF000000)>>24
+	array[ index+3 ] = (value & (size_t)0xFF000000)>>24
 
 #define WRITE_LITTLE_ENDIAN_16( value, array, index ) array[ index ] = value&0x00FF; array[ index+1 ] = (value&0xFF00)>>8
 
 namespace util
 {
-	template<typename T>
-	void xorSwap( T& x, T& y)
-	{
-		x ^= y;
-        y ^= x;
-		x ^= y;
-	}
+   template<typename T>
+   void xorSwap( T& x, T& y)
+   {
+      x ^= y;
+      y ^= x;
+      x ^= y;
+   }
 }
